+++ git/github.go
@@ -185,6 +185,7 @@ func (g *GitHub) GetArchiveLink(org, repo, tag string) (*url.URL, error) {
 		org, repo, github.Tarball, &github.RepositoryContentGetOptions{Ref: tag})
 	if err != nil {
 		if resp != nil {
+			resp.Response.Body.Close()
 			switch resp.StatusCode {
 			case http.StatusNotFound:
 				return nil, nil